requirements.txt
fastapi
pillow
python-multipart
torch
transformers
uvicorn